+++ src/main/java/ru/itis/moviehub/repositories/ActorsRepository.java
@@ -3,5 +3,5 @@ package ru.itis.moviehub.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.itis.moviehub.models.Actor;
 
-public interface ActorsRepository extends JpaRepository<Actor, Integer> {
+public interface ActorsRepository extends JpaRepository<Actor, Long> {
 }
