+++ src/main/java/ru/itis/moviehub/repositories/ProducersRepository.java
@@ -3,5 +3,5 @@ package ru.itis.moviehub.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.itis.moviehub.models.Producer;
 
-public interface ProducersRepository extends JpaRepository<Producer, Integer> {
+public interface ProducersRepository extends JpaRepository<Producer, Long> {
 }
