+++ src/main/java/ru/itis/moviehub/repositories/ScriptwritersRepository.java
@@ -3,5 +3,5 @@ package ru.itis.moviehub.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.itis.moviehub.models.Scriptwriter;
 
-public interface ScriptwritersRepository extends JpaRepository<Scriptwriter, Integer> {
+public interface ScriptwritersRepository extends JpaRepository<Scriptwriter, Long> {
 }
