+++ src/main/java/ru/itis/moviehub/services/FilmsService.java
@@ -6,7 +6,8 @@ import java.util.List;
 
 public interface FilmsService {
     List<Film> getFilms();
-    Film getConcreteFilm(Integer id);
+
+    Film getConcreteFilm(Long id);
     List<Film> getAfisha();
     List<Film> search(String name);
     List<Film> getTopFilms();
